[programs] changed defaultreactionprogram -> reactions::UncontrolledApproximation
