+++ readdy/main/model/_internal/ProgramNames.cpp
@@ -14,9 +14,10 @@ namespace readdy {
                 const std::string ProgramName<Test>::value = "Test";
                 const std::string ProgramName<AddParticle>::value = "AddParticle";
                 const std::string ProgramName<EulerBDIntegrator>::value = "Eulerian Brownian dynamics integrator";
-                const std::string ProgramName<DefaultReactionProgram>::value = "PerformReactions";
                 const std::string ProgramName<CalculateForces>::value = "Calculate forces";
                 const std::string ProgramName<UpdateNeighborList>::value = "Update neighbor list";
+
+                const std::string ProgramName<reactions::UncontrolledApproximation>::value = "Uncontrolled approximation of reactions";
             }
         }
     }
